Modify print_int_as_binary to handle 64-bit num

+++ util.h
@@ -8,6 +8,6 @@
 #include <assert.h>
 
 uint32_t get_binary_input(int size);
-void print_int_as_binary(uint32_t num, int size); 
+void print_int_as_binary(uint64_t num, int size); 
 
 #endif
